moving scipy errors from field.py to tools/error.py
