+++ parcels/field.py
@@ -1,5 +1,6 @@
 from parcels.tools.loggers import logger
 from parcels.tools.converters import unitconverters_dict, UnitConverter, Geographic, GeographicPolar
+from parcels.tools.error import FieldSamplingError, TimeExtrapolationError
 from collections import Iterable
 from py import path
 import numpy as np
@@ -13,32 +14,6 @@ from .grid import (RectilinearZGrid, RectilinearSGrid, CurvilinearZGrid,
 
 
 __all__ = ['Field', 'VectorField']
-
-
-class FieldSamplingError(RuntimeError):
-    """Utility error class to propagate erroneous field sampling"""
-
-    def __init__(self, x, y, z, field=None):
-        self.field = field
-        self.x = x
-        self.y = y
-        self.z = z
-        message = "%s sampled at (%f, %f, %f)" % (
-            field.name if field else "Field", self.x, self.y, self.z
-        )
-        super(FieldSamplingError, self).__init__(message)
-
-
-class TimeExtrapolationError(RuntimeError):
-    """Utility error class to propagate erroneous time extrapolation sampling"""
-
-    def __init__(self, time, field=None):
-        if field is not None and field.grid.time_origin and time is not None:
-            time = field.grid.time_origin + np.timedelta64(int(time), 's')
-        message = "%s sampled outside time domain at time %s." % (
-            field.name if field else "Field", time)
-        message += " Try setting allow_time_extrapolation to True"
-        super(TimeExtrapolationError, self).__init__(message)
 
 
 class Field(object):
